Now pastes into the last found window instead of one specific window
By getting all available windows, we can find the last active window (as they are ordered by their activated order).
At the moment, I'm hard-coding this to select Window 3, because Window 1 is blank (for me), Window 2 would be the Clipboard window, making Window 3 the actual last-activated window.

As it stands, this is working for me but this may need changing in the future if that 1st window thing changes.