Built-in 'Configurator' is now functional
Strings can now be edited in-app. These are immediately actioned in-app and can be pasted as soon as they are saved. They are also written to file and are functional upon re-loading the app!

I also performed a little bit of cleanup in this commit and renamed some very badly named variables.